+++ R/ms_chat_message.R
@@ -41,7 +41,7 @@
 #' @examples
 #' \dontrun{
 #'
-#' myteam <- team("my team")
+#' myteam <- get_team("my team")
 #'
 #' chan <- myteam$get_channel()
 #' msg <- chan$list_messages()[[1]]